Extract compatibility code to partial class
